Bugfix MiniMax algorithm
Now if the depth is greater than 0 it is possible to evaluate the BoardInterface
instance immediately. Whenever the BoardInterface returns true for isGameOver() method.

It is important because now the algorithm doesn't waste time and will not go deeper
into tree search for looking for the best move. Plus without this fix algorithm will
NOT work properly and returns wrong answers (look into test cases).