Make MiniMaxAlphaBeta harder to return empty Move
When given complicated board state to MiniMaxAlphaBeta class
in order to find the best move on the board it was possible for
the class to return empty move or even exceeds time limit.

Currently, the implementation still suffers from returning empty
Move but it is far less likely. The only scenario for returning
empty move is when class is created with the small timeout so in
fact it cannot even compute the first possible move.